requirements.txt
Flask==3.0.0
openai==1.12.0
python-docx==1.1.0
PyPDF2==3.0.1
